Removed CachedDfSource and move the dfs logic into the ContextIndexSearcher
